Move Changelog to Infrastructure
